Bug fixes and new strategy
Fixed a bug on strat 3 that wasn't seating tables.

Added a 5th strategy the COVID strategy

If a table arrives with a number of guests greater than the least busy servers capacity it tries to find a second least busy server. If the number of guests is greater than two servers can handle, the table abandons. If the table can be split between two servers than they are provided the option to split. On a coin flip, if the table decides to stay they are split between the two least busy servers else they abandon.